gpsd-machine-conf: allow creation of an empty package
gpsd-machine-conf is an empty recipe that does not ship any files on the
rootfs. This recipe is targeted to be extended by a bbappend that
will provide the gpsd machine configuration.

In the case where gpsd-machine-conf is included into an image, and if no
layers is providing the machine configuration, the build fails with
the following error:

	No match for argument: gpsd-machine-conf
	Error: Unable to find a match: gpsd-machine-conf

This error is because no package was created for gpsd-machine-conf
since the recipe does not install any files.

This commit allows the creation of an empty package in order to avoid
the do_rootfs failure when this package is included into an image.

Signed-off-by: Fabien Parent <fabien.parent@linaro.org>
Signed-off-by: Khem Raj <raj.khem@gmail.com>